trend-agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
bedrock-agentcore>=0.1.0
mcp>=1.0.0
httpx>=0.27.0
